@@ -67,7 +67,7 @@
       }
     },
     "build": {
-      "command": "webpack --progress --display-error-details --config ./tools/webpack/config.babel.js",
+      "command": "webpack --progress --hide-modules --display-error-details --config ./tools/webpack/config.babel.js",
       "env": {
         "NODE_ENV": "production"
       }
@@ -95,45 +95,45 @@
     }
   },
   "dependencies": {
-    "autoprefixer": "^6.6.0",
+    "autoprefixer": "^6.6.1",
     "axios": "^0.15.3",
     "babel-core": "^6.21.0",
     "babel-runtime": "^6.20.0",
-    "better-npm-run": "0.0.13",
+    "better-npm-run": "0.0.14",
     "chalk": "^1.1.3",
     "compression": "^1.6.2",
     "express": "^4.14.0",
     "extract-text-webpack-plugin": "^2.0.0-beta.3",
-    "helmet": "^3.2.0",
+    "helmet": "^3.4.0",
     "hpp": "^0.2.1",
     "immutable": "^3.8.1",
     "morgan": "^1.7.0",
     "opn": "^4.0.2",
-    "react": "^15.4.1",
-    "react-addons-shallow-compare": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "react": "^15.4.2",
+    "react-addons-shallow-compare": "^15.4.2",
+    "react-dom": "^15.4.2",
     "react-helmet": "^3.3.0",
     "react-hot-loader": "^3.0.0-beta.6",
     "react-immutable-proptypes": "^2.1.0",
-    "react-redux": "^5.0.1",
-    "react-router": "^3.0.0",
+    "react-redux": "^5.0.2",
+    "react-router": "^3.0.1",
     "react-router-redux": "^4.0.7",
     "redbox-react": "^1.3.3",
     "redux": "^3.6.0",
-    "redux-immutable": "^3.0.9",
+    "redux-immutable": "^3.0.10",
     "redux-thunk": "^2.1.0",
     "serialize-javascript": "^1.3.0",
     "serve-favicon": "^2.3.2",
-    "stylelint-webpack-plugin": "^0.4.2",
-    "webpack": "^2.2.0-rc.3",
+    "stylelint-webpack-plugin": "^0.5.1",
+    "webpack": "^2.2.0-rc.4",
     "webpack-dev-middleware": "^1.9.0",
-    "webpack-hot-middleware": "^2.14.0",
-    "webpack-isomorphic-tools": "^2.6.5"
+    "webpack-hot-middleware": "^2.15.0",
+    "webpack-isomorphic-tools": "^2.6.6"
   },
   "devDependencies": {
     "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.10",
-    "babel-plugin-istanbul": "^3.0.0",
+    "babel-plugin-istanbul": "^3.1.2",
     "babel-plugin-system-import-transformer": "^2.4.0",
     "babel-plugin-transform-runtime": "^6.15.0",
     "babel-preset-latest": "^6.16.0",
@@ -144,12 +144,12 @@
     "coveralls": "^2.11.15",
     "css-loader": "^0.26.1",
     "enzyme": "^2.7.0",
-    "eslint": "^3.12.2",
-    "eslint-config-airbnb": "^13.0.0",
+    "eslint": "^3.13.1",
+    "eslint-config-airbnb": "^14.0.0",
     "eslint-loader": "^1.6.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.8.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
+    "eslint-plugin-react": "^6.9.0",
     "image-webpack-loader": "^3.1.0",
     "imports-loader": "^0.7.0",
     "json-loader": "^0.5.4",
@@ -160,22 +160,22 @@
     "karma-mocha-reporter": "^2.2.1",
     "karma-phantomjs-launcher": "^1.0.2",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.8.1",
+    "karma-webpack": "^2.0.1",
     "mocha": "^3.2.0",
-    "node-sass": "^4.1.1",
+    "node-sass": "^4.3.0",
     "nodemon": "^1.11.0",
     "null-loader": "^0.1.1",
     "phantomjs": "^2.1.7",
-    "postcss": "^5.2.8",
-    "postcss-loader": "^1.2.1",
-    "react-addons-test-utils": "^15.4.1",
+    "postcss": "^5.2.10",
+    "postcss-loader": "^1.2.2",
+    "react-addons-test-utils": "^15.4.2",
     "react-hot-loader": "^3.0.0-beta.6",
     "redux-mock-store": "^1.2.1",
     "rimraf": "^2.5.4",
     "sass-loader": "^4.1.1",
     "sinon": "^2.0.0-pre.4",
     "style-loader": "^0.13.1",
-    "stylelint": "^7.7.0",
+    "stylelint": "^7.7.1",
     "stylelint-config-standard": "^15.0.1",
     "url-loader": "^0.5.7"
   }
